Algo test

+++ include/lem_in_stacks.h
@@ -6,7 +6,7 @@
 /*   By: thflahau <thflahau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/18 16:03:49 by thflahau          #+#    #+#             */
-/*   Updated: 2019/04/23 13:42:25 by thflahau         ###   ########.fr       */
+/*   Updated: 2019/04/24 16:49:43 by thflahau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,15 +15,15 @@
 
 # include <lem_in_queue.h>
 
-typedef struct			s_stack
+typedef struct				s_stack
 {
-	uint64_t			size;
-	struct s_queue		*path;
-	struct s_stack		*next;
-}						t_stack;
+	uint64_t				size;
+	struct s_queue			*path;
+	struct s_stack			*next;
+}							t_stack;
 
 void						ft_free_stacks(t_stack **head);
 void						ft_print_movements(t_map *map, t_stack *list);
-uint8_t						ft_push_path_to_stack(t_map *map, t_stack **stack);
+uint64_t					ft_push_path_to_stack(t_map *map, t_stack **stack);
 
 #endif
