+++ srcs/queue.c
@@ -6,14 +6,14 @@
 /*   By: thflahau <thflahau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/16 11:50:32 by thflahau          #+#    #+#             */
-/*   Updated: 2019/04/23 13:47:14 by thflahau         ###   ########.fr       */
+/*   Updated: 2019/04/24 16:19:57 by thflahau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include <lem_in_queue.h>
 
-static inline t_queue			*ft_allocate_queue_node(uint64_t key)
+static inline t_queue			*ft_allocate_queue_node(uint32_t key)
 {
 	t_queue						*node;
